now returns user's tasks instead of all

+++ internal/adapter/repo/postgres/tasks.go
@@ -11,19 +11,19 @@ import (
 	"github.com/Emin-07/TaskManager/internal/core/domain"
 )
 
-func (m TaskRepo) List(ctx context.Context, limit, offset int) ([]*repo.TaskDb, error) {
+func (m TaskRepo) List(ctx context.Context, limit, offset, userId int) ([]*repo.TaskDb, error) {
 	var tasks []*repo.TaskDb
-	query := `SELECT * FROM tasks WHERE expires > CURRENT_TIMESTAMP  ORDER BY id LIMIT $1 OFFSET $2`
-	err := m.DB.SelectContext(ctx, &tasks, query, limit, offset)
+	query := `SELECT * FROM tasks WHERE expires > CURRENT_TIMESTAMP AND user_id = $1 ORDER BY id LIMIT $2 OFFSET $3`
+	err := m.DB.SelectContext(ctx, &tasks, query, userId, limit, offset)
 	if err != nil {
 		return nil, err
 	}
 	return tasks, nil
 }
 
-func (m TaskRepo) Get(ctx context.Context, id int) (*repo.TaskDb, error) {
+func (m TaskRepo) Get(ctx context.Context, id, userId int) (*repo.TaskDb, error) {
 	task := repo.TaskDb{}
-	err := m.DB.GetContext(ctx, &task, "SELECT * FROM tasks WHERE id = $1", id)
+	err := m.DB.GetContext(ctx, &task, "SELECT * FROM tasks WHERE id = $1 AND user_id = $2", id, userId)
 	if err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
 			return nil, domain.ErrNoRecord
@@ -56,7 +56,7 @@ func queryOrderTracker(query *strings.Builder, isNotFirst *bool) {
 	}
 }
 
-func (m TaskRepo) Patch(ctx context.Context, title, text string, priority, expireDays, id int) error {
+func (m TaskRepo) Patch(ctx context.Context, title, text string, priority, expireDays, id, userId int) error {
 	var query strings.Builder
 	var args []any
 	var isNotFirst bool
@@ -90,7 +90,8 @@ func (m TaskRepo) Patch(ctx context.Context, title, text string, priority, expir
 		return domain.ErrNoData
 	}
 	args = append(args, id)
-	query.WriteString(fmt.Sprintf("WHERE id = $%d", cnt))
+	args = append(args, userId)
+	query.WriteString(fmt.Sprintf("WHERE id = $%d AND user_id = $%d", cnt, cnt+1))
 
 	result, err := m.DB.ExecContext(ctx, query.String(), args...)
 
@@ -106,8 +107,8 @@ func (m TaskRepo) Patch(ctx context.Context, title, text string, priority, expir
 	return nil
 }
 
-func (m TaskRepo) Delete(ctx context.Context, id int) error {
-	_, err := m.DB.ExecContext(ctx, "DELETE FROM tasks WHERE id = $1", id)
+func (m TaskRepo) Delete(ctx context.Context, id, userId int) error {
+	_, err := m.DB.ExecContext(ctx, "DELETE FROM tasks WHERE id = $1 AND user_id = $2", id, userId)
 	if err != nil {
 		return err
 	}
